requirements.txt
dnspython==1.16.0
inflection==0.3.1
pymongo==3.8.0
apache-airflow-providers-mongo==1.0.1
